Move URL node extraction from util to the config package
